pyproject.toml
[project]
name = "flight-delay-gnn-system"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = "==3.11.9"
dependencies = [
    "cassandra-driver>=3.29.3",
    "findspark>=2.0.1",
    "kafka-python>=2.3.0",
    "matplotlib>=3.10.8",
    "meteostat>=1.7.6",
    "pandas>=2.3.3",
    "pyarrow>=22.0.0",
    "pyspark==3.5.1",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.8.0",
    "streamlit>=1.52.2",
    "torch>=2.9.1",
    "torch-geometric>=2.7.0",
]
